+++ Menus/MainMenuGUI.cpp
@@ -222,9 +222,8 @@ namespace RTE {
 
 		std::string metaNotice = {
 			"- A T T E N T I O N -\n\n"
-			"Please note that the Campaign is in an INCOMPLETE, fully playable, yet still imperfect state!\n"
-			"As such, it is lacking some polish, audio, and game balancing, and we will be upgrading it significantly in future.\n"
-			"That said, you can absolutely enjoy fighting the A.I. and/or up to three friends in co-op, 2 vs 2, etc.\n\n"
+			"Please note that Conquest mode is very INCOMPLETE and flawed, and the Cortex Command Community Project team will be remaking it from the ground-up in future. "
+			"For now though, it's fully playable, and you can absolutely enjoy playing it with the A.I. and/or up to three friends. Alternatively, check out Void Wanderers on cccp.mod.io!\n\n"
 			"Also, if you have not yet played Cortex Command, we recommend you first try the tutorial:"
 		};
 		metaNoticeLabel->SetText(metaNotice);